Updated to allow self changeing username

+++ events.go
@@ -28,7 +28,7 @@ type RateLimit struct {
 // Event provides a basic initial struct for all websocket events.
 type Event struct {
 	Operation int32             `json:"op"`
-	Sequence  int32             `json:"s"`
+	Sequence  int64             `json:"s"`
 	Type      string          `json:"t"`
 	RawData   json.RawMessage `json:"d"`
 	// Struct contains one of the other types in this file.
